requirements.txt
pydantic>=2.0.0
pydantic-ai>=0.0.14
PyMuPDF>=1.23.0
python-pptx>=0.6.21
Pillow>=10.0.0
anthropic>=0.40.0
